move the mockdata to new file
