+++ src/features/grading/writingmockdata.jsx
@@ -1,0 +1,49 @@
+const mockData = {
+  part1: {
+    instructions:
+      "Part 1. You want to join the Museum club. You have 5 messages from a member of the club. Write short answers (1-5 words) to each message. Recommended time: 3 minutes. (1 point)\n* (You're allowed to write up to 10 words without affecting your grade).",
+    questions: [
+      'Are you interested in art?',
+      'Where did you go last night?',
+      'What are you wearing today?',
+      'Which sport is the most popular in your country?',
+      'What do you like doing with your friend?'
+    ],
+    answers: [
+      'Yes, I love paintings. They are beautiful and inspiring. I often visit art galleries. It makes me feel creative.',
+      'I went to a café. I wanted to relax. I enjoyed a cup of coffee. It was a peaceful night.',
+      "A white shirt and jeans. It's comfortable and stylish. I feel confident wearing it. I like simple outfits.",
+      'Football is the most popular. Many people watch and play it. We have big football tournaments. It brings people together.',
+      ''
+    ]
+  },
+  part2: {
+    instructions: 'Part 2. Write a short essay about your favorite museum. (5 points)',
+    questions: ['Describe your favorite museum and why you like it.'],
+    answers: [
+      'The National Art Museum is my favorite. It has an impressive collection of modern and classical art. The architecture of the building itself is a masterpiece with its grand columns and spacious galleries. I particularly enjoy the interactive exhibits that allow visitors to engage with the art in unique ways. Every time I visit, I discover something new and fascinating.'
+    ]
+  },
+  part3: {
+    instructions: 'Part 3. Compare and contrast two different types of museums. (10 points)',
+    questions: ['Compare art museums and science museums.', 'Discuss the educational value of each type.'],
+    answers: [
+      'Art museums showcase human creativity through paintings, sculptures, and other visual media, while science museums focus on natural phenomena and technological innovations through interactive exhibits. Art museums often have a contemplative atmosphere, whereas science museums encourage hands-on exploration.',
+      'Both types offer significant educational value. Art museums cultivate aesthetic appreciation and cultural understanding, helping visitors develop critical thinking about visual communication. Science museums, on the other hand, promote scientific literacy and curiosity about how the world works through experimentation and demonstration.'
+    ]
+  },
+  part4: {
+    instructions: 'Part 4. Analyze the role of museums in modern society. (15 points)',
+    questions: [
+      'How have museums evolved in the digital age?',
+      'What is the social impact of museums in communities?',
+      'Lmao i just delete wrong'
+    ],
+    answers: [
+      'Museums have embraced digital technology by creating virtual tours, interactive displays, and online collections. Many now offer mobile apps that enhance the visitor experience with additional information and augmented reality features. Social media has also become a vital tool for museums to engage with audiences beyond their physical locations.',
+      '',
+      'Future museums will likely become more interactive and personalized, using AI to tailor experiences to individual interests. We may see more community-curated exhibits and collaborative spaces that blur the line between creator and visitor. Sustainability will become central to museum operations, with eco-friendly practices integrated into both building design and exhibition planning.'
+    ]
+  }
+}
+export default mockData
